pyproject.toml
[project]
name = "groupproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pillow>=12.1.1",
    "pyarrow>=23.0.1",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
]
